- add apps/{khtml,khtml/css} - check for packaged files from FHS.spec
Changed files:
    kde-common-dirs.spec -> 1.7

+++ kde-common-dirs.spec
@@ -1,8 +1,11 @@
+# avoid rpm 4.4.9 adding rm -rf buildroot, we need the dirs to check consistency
+%define		__spec_clean_body	%{nil}
+%define		_enable_debug_packages	0
 Summary:	K Desktop Environment - common directories
 Summary(pl.UTF-8):	Wspólne katalogi KDE (K Desktop Environment)
 Name:		kde-common-dirs
 Version:	0.1
-Release:	1
+Release:	2
 License:	LGPL
 Group:		X11/Libraries
 URL:		http://www.kde.org/
@@ -38,7 +41,22 @@ install -d \
 	$RPM_BUILD_ROOT%{_desktopdir}/kde
 
 %clean
-rm -rf $RPM_BUILD_ROOT
+cd $RPM_BUILD_ROOT
+check_filesystem_dirs() {
+	RPMFILE=%{name}-%{version}-%{release}.%{_target_cpu}.rpm
+	TMPFILE=$(mktemp)
+	find | sed -e 's|^\.||g' -e 's|^$||g' | LC_ALL=C sort > $TMPFILE
+
+	# find finds also '.', so use option -B for diff
+	if rpm -qpl %{_rpmdir}/$RPMFILE | grep -v '^/$' | LC_ALL=C sort | diff -uB $TMPFILE - ; then
+		rm -rf $RPM_BUILD_ROOT
+	else
+		echo -e "\nNot so good, some directories are not included in package\n"
+		exit 1
+	fi
+	rm -f $TMPFILE
+}
+check_filesystem_dirs
 
 %files
 %defattr(644,root,root,755)
@@ -47,6 +65,8 @@ rm -rf $RPM_BUILD_ROOT
 %dir %{_datadir}/applnk/.hidden
 %dir %{_datadir}/apps
 %dir %{_datadir}/apps/kde
+%dir %{_datadir}/apps/khtml
+%dir %{_datadir}/apps/khtml/css
 %dir %{_datadir}/apps/kconf_update
 %dir %{_datadir}/apps/kstyle
 %dir %{_datadir}/apps/kstyle/themes
